Moved scaffolds directory. Added bottom_navigation_bar and toolbar directories.
